move source files to root, change where port number is used
